Tiro com mouse e teclado funcionando, com mouse atira no ângulo correto.

+++ MyZelda/src/me/rlbpc/entities/BulletShoot.java
@@ -14,10 +14,10 @@ import me.rlbpc.world.Camera;
  *
  */
 public class BulletShoot extends Entity{
-	private int dx, dy;
+	private double dx, dy;
 	private double bulletSpeed = 4;
 
-	public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, int dx, int dy) {
+	public BulletShoot(int x, int y, int width, int height, BufferedImage sprite, double dx, double dy) {
 		super(x, y, width, height, sprite);
 		this.dx = dx;
 		this.dy = dy;
